add mongo storage layer

+++ models/student.go
@@ -4,5 +4,5 @@ type Student struct {
 	ID    string  `json:"id"`
 	Name  string  `json:"name"`
 	Email string  `json:"email"`
-	GPA   float32 `json:"gpa"`
+	GPA   float64 `json:"gpa"`
 }
